refactor: only use certified GitHub Actions to increase security
Move the spell checker Action to the lint file.  Less files so easier to maintain.

Add a Dependabot config file and configure checks for updates to GitHub Actions everyday.

`actions/setup-python@v2` is a certified GitHub Action and so is `actions/setup-node@v1` but `ibiqlik/action-yamllint@v1` is a 3rd party library.

Official -> "Using third-party actions
The individual jobs in a workflow can interact with (and compromise) other jobs. For example, a job querying the environment variables used by a later job, writing files to a shared directory that a later job processes, or even more directly by interacting with the Docker socket and inspecting other running containers and executing commands in them.

This means that a compromise of a single action within a workflow can be very significant, as that compromised action would have access to all secrets configured on your repository, and can use the GITHUB_TOKEN to write to the repository. Consequently, there is significant risk in sourcing actions from third-party repositories on GitHub. "

https://docs.github.com/en/actions/learn-github-actions/security-hardening-for-github-actions#using-third-party-actions
https://docs.github.com/en/github/administering-a-repository/configuration-options-for-dependency-updates
https://github.com/marketplace/actions/setup-python
https://github.com/marketplace/actions/setup-node-js-environment
https://github.com/marketplace/actions/yaml-lint