Update guide to install options on Mac
The previous version was missing a couple of important alternatives. (I could not find better "guides" for those alternatives, here in the docs. So I offered the only suitable resource links I I could find. Is there in fact a "guide" for running the Express edition, or for he docker images?) 

Also tweaked some grammar/typos.